EICNET-2758: Re-factor code after implementation of eic_moderation.helper service + code clean-up.
